Improved code, index.php
